refactor: remove 0x utils
