+++ src/quote/leveraged-extended/provider.ts
@@ -7,7 +7,7 @@ import {
   LeveragedTokenData,
 } from 'utils/leveraged-token-data'
 import { slippageAdjustedTokenAmount } from 'utils/slippage'
-import { Exchange, SwapData } from 'utils/swapData'
+import { Exchange, SwapData } from 'utils'
 
 import { QuoteProvider } from '../quoteProvider'
 import { QuoteToken } from '../quoteToken'
