+++ src/utils/index.ts
@@ -1,9 +1,8 @@
-export * from './0x'
 export * from './contracts'
 export * from './issuanceModules'
 export * from './numbers'
 export * from './overrides'
 export * from './slippage'
-export * from './swapData'
+export * from './swap-data'
 export * from './tokens'
 export * from './UniswapPath'
